Rename causal disentangled representation learning to CausalVAE.
